+++ WindowsFormsSau/WindowsFormsSau.csproj
@@ -54,11 +54,11 @@
     <Compile Include="FormParking.Designer.cs">
       <DependentUpon>FormParking.cs</DependentUpon>
     </Compile>
-    <Compile Include="FormSau.cs">
+    <Compile Include="FormVehicle.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FormSau.Designer.cs">
-      <DependentUpon>FormSau.cs</DependentUpon>
+    <Compile Include="FormVehicle.Designer.cs">
+      <DependentUpon>FormVehicle.cs</DependentUpon>
     </Compile>
     <Compile Include="ITransport.cs" />
     <Compile Include="Parking.cs" />
@@ -70,8 +70,8 @@
     <EmbeddedResource Include="FormParking.resx">
       <DependentUpon>FormParking.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="FormSau.resx">
-      <DependentUpon>FormSau.cs</DependentUpon>
+    <EmbeddedResource Include="FormVehicle.resx">
+      <DependentUpon>FormVehicle.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
